[util] Fixing the behavior of the cleanFolder function in Files class.
The previous behavior does not remove the parent folder when its has no
file inside and the "deleteParentFolder" argument is true.
The behavior of the function is updated to delete the parent folder as
expected.
Additionally, the recurive algorithm is replaced by an iterative
algorithm in order to make it more robust.

Signed-off-by: Stéphane Galland <galland@arakhne.org>