[releng] Add job for updating versions
This job checks out all repositories and creates a target branch on them
to apply version changes. The job can update the Xtext version to
manifests and POMs, and it can set the Xtext bootstrap version. Which
update is performed depends on the given parameters. The job will
usually perform only one update.

The changes are pushed and a pull request is opened for the changes.

Signed-off-by: Karsten Thoms <karsten.thoms@itemis.de>