[#919] Add quickfixes that adds JUnit 4 or 5 libs to the classpath
Similar to the quickfix that add Xtend libs, it will first check if the
project is a plugin project and add them to the MANIFEST.MF, if not it
will add entries to the .classpath. The quickfixes appear for errors on
import declaration that starts by "org.junit" or
"org.junit.jupiter".

Signed-off-by: Vivien Jovet <vivien.jovet@gmail.com>