[eclipse/xtext#1455] Fixing CompletionTest.
The tests work with concrete offsets and hence can only work with '\n'
because '\r\n' will always lead to broken expectations. Test case in
CompletionTest is only broken because TestLanguageProposalCreator, a
class used only for testing, adds 1 to the offset but "2" would be
needed for input with '\r\n'. In other places it is already ensured that
'\n' is used (e.g. AbstractLanguageServerTest#assertEquals). Hence just
align the input text to the rest of the implementation and using '\n'.

Signed-off-by: Arne Deutsch <Arne.Deutsch@itemis.de>