[eclipse/xtext-xtend#903] Fix @Accessors not generating `@Deprecated` or `@Override`
Now `@Accessors` generates `@Override` when required. It is still an error to attempt to override a final method.
Now the annotation detects deprecated fields and generates `@Deprecated` when required. This behaviour can be changed via `Accessors.deprecationPolicy`:
`Accessors` can now be instructed to deprecate either generated method (or both or none), or to mirror the annotated field's deprecation (this is the default).

Signed-off-by: ivaniesta14 <ivaniesta14@gmail.com>