[#1466] Global unique name validator (#1483)
This change introduces an explicit validation context that is used by the INamesAreUniqueValidationHelper. This allows to fine tune the scope of the validation in various ways. Default contexts are available to validation uniqueness in the current container, along the chain of visible containers or local to the current resource (the default).

Local unique name validation can be implemented now based on the LocalUniqueNameContext.

closes #1466