[bugzilla 564578] Fix wrong colors on colors preference page
This fixes an issue on the syntax coloring preference page:
when one color is changed, all other colors were shown as black

The preference store of field editors should not be replaced when
when a color is changed. The preference store is already a chained
preference store that contains the internal store.

This fixes bugzilla 564578

Signed-off-by: Hendrik Eeckhaut <hendrik.eeckhaut@sigasi.com>