[#1078] Register theming in XtendHighlightingConfiguration
Avoid the need to initialize theming in generated activator.
Preferences are only already set when switching to dark mode. When
switching back to default this does not work somehow. As a workaround
preferences are reset with the highlighting configuration.

Added the default theme anyway. Might be useful later.

Configured access rules as ThemeEngine and ITheme are internal API.

Signed-off-by: Karsten Thoms <karsten.thoms@karakun.com>