Fix NPE if there is no resource provider for an URI
Fix NPE if there is no resource provider for an URI by checking for null
and defaulting to false if null, the null check is equivalent to the one
already present in getDeltasForDeletedResources. In addition mark the
URI for removal.

The resource provider can be null in our development environments or if
a persisted index is being reused after some language was de-installed.