waitForBuild throws with reference to original exception
Currently, waitForBuild catches a CoreException and throws an
OperationCanceledException with just the message of the original
CoreException. Indeed, OperationCanceledException does not have a
constructor for passing also an exception. Without the original
exception, though, it is hard to understand what went wrong during the
test, since we lost the original exception's stacktracke

Signed-off-by: Lorenzo Bettini <lorenzo.bettini@gmail.com>