Add API for textual Code Actions
In addition, improve naming of variables in the existing API to use
diagnostics for org.eclipse.lsp4j.Diagnostic instead of issues and
errors, and resolutions for DiagnosticResolution instead of issues in
the DiagnosticResolutionAcceptor.

A new class TextEditorAcceptor has been factored out of ChangeConverter2

The signature of addTextEdit has been adapted to conform to the Xtext
coding standards (an acceptor normally do not return a value).

Signed-off-by: Rubén Porras Campo <Ruben.PorrasCampo@avaloq.com>