Storing a binary resource removes its URI from the SourceLevelURIs set
by doing that we increase the performance of the
BuildContext#executeClustered because it will be possible to use binary
models from resources which has already been built and stored.

The methods to deal with binary models and source lever uris have been
streamlined and moved to the IncrementalBuilder class. A new method
deleteBinaryResources has been added for downstream projects which
support deletion of binary resources.

Signed-off-by: rubenporras <Ruben.PorrasCampo@avaloq.com>