Make Xtext compatible with Guava 31
Widen the version ranges in all MANIFEST.MF where the guava bundle is
required to "[30.1.0,32.0.0)" and test Xtext with Guava 31 in the
xtext-latest.target.
Guice still requires Guava 30 so that version is still required, but
this change should at least allow to use Xtext in conjunction with Guava
31.1. At the moment that is very problematic because Xtext plugins
reexport the guava bundle.
Because Guice uses Guava only internally and not in its API Guava
classes of version 31 cannot 'leak' into Xtext and class-space
consistency is ensured.