[Setup] Reference xtext target-files in Xtext's Oomph Setup directly
This avoids the need to replicate and maintain all different target
platforms as pure Oomph Targlets.
Use one TargletTask for each xtext-target file (the targlet-task
contains the target-file as 'composedTarget') and use a corresponding
filter to activate that task if the 'eclipse.target.platform' variable
has a corresponding value.

Only for the o.e.cbi.p2repo.analyzers dependencies a always active
TargletTask with corresponding Targlet is created, because they are not
contained in the target-definition specified by the target-files.
That targlet is merged with the one activated for the selected Eclipse
target. Without the p2repo-analyzers dependencies or if they where
contained in the target-files Xtext could use a simple
TargetPlatformTask (if it is not desired to merge the TP with the
Targlets defined by other Projects in the workspace).