Remove 'de.itemis.xtext.antlr.feature' from target files
That feature is only used for developer convenience in the Xtext
development workspace and already added to that by the Xtext
Oomph-setup.
Use 'DynamicImport-Package' instead of Required-Bundle (or
Import-Package) to make Xtext still compile without that dependency.
Since the classes form 'de.itemis.xtext.antlr' are accessed reflectively
they are not needed at compile-time.