Support offloading the node-model while a resource is loaded
The node model consumes a lot of memory compared to the resource
contents. Languages may opt-in to offload the node model into a
serialized byte[] either in-memory or persisted somewhere else. It is
re-attached on demand.

There are some caveats, though. No strong references to original nodes
must be held. Otherwise the node model cannot be detached. If a code
path strictly needs to refer to a node,
DetachableParseResultWrapper.externalReference(IParseResult, INode) can
be used.

Feature can be enabled by using the DetachableNodeModelFragment. It is
subject to the language to chose a suitable moment in the life-cycle of
a resource to schedule unloading the node model.
DetableParseResultWrapper.acquire might be a candidate if all resources
shall be marked as unloadable immediately.

Also:
- Improved testability of node model
- New more flexible API for node model serialization/deserialization
- Allow to unbind a Guice binding from a generator fragment

Signed-off-by: Sebastian Zarnekow <sebastian.zarnekow@gmail.com>